Added basic footer html file
This is just the basic set up for the footer. I will be working on improving this, just wanted to have some form of it for testing purposes. This first upload includes four blank footer links, a temporary link to the admin page (as per Joey's request) and three spots for social media links (as of right now, just basic styled img objects, will be changed down the line).